adding new UI system that links with players
This ui system checks the number of devices and select ui according to
the number of devices.A gameobject (canvas) called gameUI (soon to be
saved as a prefab) has 4 gameobjects for 1 to 4 players.
Each game object has a number of pannels sized and placed according to
the number of players.
The UIManager script will automatically choose the right UI 0.1s after
starting the scene (coroutine) and it will also automatically each
pannel to a player.

-- Next step --
Now that the player is retrieved by a pannel, we need to use it to get
useful information that could be important to the player.

To do :
- Choose the info to display and how.
- Place the elements.
- Inject information to the UI Elements (in the update method)